requirements.txt
aniso8601==8.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
FHIR-Parser==0.1.5
Flask==1.1.1
Flask-RESTful==0.3.8
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
lxml==4.5.0
MarkupSafe==1.1.1
numpy==1.18.2
pandas==1.0.3
plotly==4.5.4
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2019.3
requests==2.23.0
retrying==1.3.3
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.0
